Added main functions of the package
This is file containing the main functions from the package. The functions were tested on four different test sets depending on the type of function calculating lag. The variables in the package have been simplified and unified to a standard consistent with the guide of Advanced R by Hadley Wickham.